Garbage collection improvement, force app update prompt, more...
Some Notable Changes:
- Adjusted garbage collection call to execute when the window count
reaches zero. It previously was called manually when the Clear Windows
button was pressed, but that didn't cover cases where each window was
closed individually. In those cases, the memory would sometimes linger
at high amounts on some systems.
- Added option to force the app update prompt when updating manually via
the button within the configuration window. The primary use case for
this is reinstalling the latest version in the selected branch, even if
it matches the current version. This not only serves for
troubleshooting, but also updating to new builds of the same version and
branch. Think SNAPSHOT builds for example, which will be updated
multiple times before moving to BETA and onwards.
- Removed old code and documentation.