First stage of fullscreen improvements.
- Entering fullscreen mode will now hide the window borders, allowing
for edge-to-edge display of media.
- Videos with an aspect ratio that does not match the display will now
tend to fill in the transparent space with a black background for better
immersion. This was a natural consequence of changing the borders and is
not final. It may be changed back in the future or become a
configuration option.
- Improved the window state system and created recurring and permanent
hooks. Recurring hooks are not disposed of after their first execution.
Permanent hooks do the same and also cannot be unhooked. They are
forever attached to that window state object, and therefore the window.
- Moved and unified fullscreen logic to hooks on the new FULLSCREEN
window state property.