Drag & Drop Rewrite, fullscreen improvement, binary updates, more...
Some Notable Changes:
- Reworked drag and drop media "flavor" picking process and brought
meaning to the word 'Prefer' in the configuration option: Prefer Links
with Drag and Drop. PiPAA windows will now attempt to load the media as
another flavor if the first one fails. Before, loading a bad link from a
drag and drop would fail the entire loading process. Now, if a valid
file or image is also retrievable from the drag and drop, PiPAA will
attempt to load that next.
- A window can no longer be dragged/moved while in fullscreen mode.
- Added new window state property for when media is being closed.
- Reduced code redundancy and improved readability in a few classes,
primarily with methods pertaining to the drag and drop functionality.
- Updated yt-dlp binary stable@2024-08-06 -> stable@2024-11-18
- Updated gallery-dl binary 1.27.2 -> 1.27.7