Added support for MPEG-DASH (mpd) media. Rare bug fixes.
Some Notable Changes:
- Added support for MPEG-DASH (mpd) media. MPD media can be buggy to
play and slow to cache, mainly due to internal VLC errors.
- Fixed a potential bug which could occur when trying to load direct web
media of certain supported file extensions. PiPAA may have incorrectly
said that the media was unsupported, causing the loading to fail.
- Fixed a bug that could occur when caching PLAYLIST media. The
PLAYLIST's extension would be used in the file name instead of a video
extension. This obviously leads to inconsistencies, as downloaded
content is not another PLAYLIST, but the actual media the PLAYLIST
refers to.
- Minor documentation improvements.