Fixed attribution bug caused by dc70a98 which led to loading failures.
Noticed unintentional inversion of logic in an if statement from commit
dc70a98. This was causing the attributor to pick the wrong media
extension, which then became misaligned with the expected extension
during the download process. The download would ultimately succeed, but
the name would often wrongly include two extensions, and the user would
be told that the media was invalid when it was not. This has been fixed.