Improvements to saving audio artwork, fixed a fullscreen bug, and more.
Some Notable Changes:
- Saving artwork to media will now perform a custom reload, keeping
things going from where they left off while still displaying the new
artwork. Previously, the whole thing was reloaded, resetting playback
position, window size, and more. That was annoying.
- The user will now be notified if there was an internal error while
attempting to save the new artwork to the media.
- Fixed a bug causing window sizing glitches when entering or exiting
fullscreen with audio media playing in the window. The bug may have to
do with the fact that the VLC player handling the audio playback does
not have any visible content to display, plus it is positioned
differently in the window. In that case, using its fullscreen
implementation causes sporadic sizing glitches. It seems to consistently
change size properly when using the Swing approach, as used for standard
image media.
- Reworked media command arguments system and created selection
classes/system.