Attribution/Remote Media fixes and small quality of life improvements.
Some Notable Changes:
- Opening media file location now selects the file within the shown
folder.
- Adjusted functionality and visuals of always-on-top message dialogs,
adding more padding and allowing for multi-line messages.
- Improved error logic while setting remote media: The user is now
notified if a download failed strictly because of their Download Media
configuration.
- The window border now flashes red to indicate an error if the process
of setting remote media returned a null result.
- Improved overall speed of the downloading/caching process by first
trying to use the same binary which succeeded during attribution.
- Fixed a bug where some sites would fail to attribute and/or load due
to the cookies argument passed to yt-dlp or gallery-dl. The "--cookies"
argument allows the user to play/cache media which requires credentials.
However, some sites do not seem to appreciate receiving this argument if
no credentials exist for that site in the cookies file. This results in
those sites breaking entirely. The attributor now figures out if the
site supports cookies and passes that information on to the window for
when it tries to set the remote media. The attributor will prefer
cookies, but will fall back on not using them if necessary.
- Minor code and documentation adjustments, fixes, and optimizations.