Significantly improved accuracy and feel of zoom actions. Zoom bug fix.
- Fixed a bug which caused zoom actions to skew towards the direction of
a previous zoom, despite starting from an unzoomed state.

Technical Description of Changes:
Previous algorithm simply took the zoom point and capped its effect
within the bounds of an arbitrary constant. With such a bad solution,
and with the constant being so small, this led to a zoom that was very
inconsistent and hard to control at most times. The pan actions would
heavily favor one of the four corners, and zooming towards the center
when not exactly over it could be challenging.

The new algorithm is linear across the entire area of the display
component which holds the image. At small window sizes, the pan factor
constant had too great an effect, so a power function was tested.
However, while it did work, it could make the zoom pan feel inaccurate.
Instead, the pan factor now scales with the greatest dimension (x/y) of
the component. Pan zoom actions move the image less in pixels when the
window is smaller, and move the image more when the window is larger.
This works well because the rendered image itself has less pixels when
the window is smaller, making each moved pixel look like a much greater
movement.