Increased delay for fullscreen shortcut using mouse double-click.
The default delay was a little too generous. It could trigger when
quickly panning an image/GIF multiple times in a row while in the Swing
viewer. Increasing the default delay for now so fullscreen attempts must
be more intentional. Users can customize this control and delay once
shortcuts customization is fully implemented.