Fixed a bug that caused inconsistent minimums when resizing a window.
With this bug, resizing a window was not consistent across the different
directions. If the media was wide, resizing all the way down from the
West or East would result in the window's height being below the
minimum. Resizing from the North or South would similarly result in the
window's width being below the minimum.

This was caused by the resizing code not accounting for the effect it
had on the other dimension. When displaying media and considering the
aspect ratio, the code will now calculate ratio-preserving minimums to
ensure that neither dimension becomes too small.