Fixed a rare bug where new windows would be invisible and unusable.
The bug seemed most common on application launch with the default, empty
window. This almost certainly stemmed from said windows being created
off of Swing's event-dispatch thread (EDT), which is critical and
required to be used for GUI changes. This, of course, includes creating
new windows. Other parts of the app respect this, but some crucial areas
were missed or have yet to be addressed.