Differentiated between Normal and Strict cache outputs for Trim option.
Small quality of life improvement. For example, on media with a
transparent background and soft drop shadow, the Normal mode can trim
some of the outermost, partially transparent pixels. After switching to
Strict mode, the cached output from Normal mode would still be loaded,
since the output wasn't separated and was saved under the same name
extension in the cache. This has been changed so that they are trimmed
and cached under unique names. Users will not have to clear the cache,
or find and delete the file, just to see a change in switching trimming
modes.