Dynamically adjust minimum window size depending on current media.
Previously, the window would enforce a 64x64 content minimum when
resizing a window. That minimum will now adapt to a smaller amount to
accommodate the current media. This allows for incredibly small media to
be displayed, without issue, at its native size.

There is currently a new, but small, absolute minimum content size of
4x4 pixels. This will likely change in the future and function
differently than it does now.