Add set of custom throwing function interfaces.
The interfaces all throw Exception, may or may not return a value, and
take up to five arguments. This provides a lot of flexibility and
options depending on the needs of the code.
Lambda implementations can utilize code that throws exceptions without
the need of an inner try/catch block. Exceptions can be caught outside.
These interfaces may or may not be needed in the future.