Thread monitoring, process execution updates, and major known issue fix.
- Implemented Thread monitoring system.
Asynchronous executions can now be monitored and easily interrupted at
any point. This is especially useful for process executions, which would
otherwise continue running until they completed or the entire
application closed. Even then, child processes would continue running in
the background.
- Fixed known issue where background processes would run past app exit.
As described above, this was addressed by monitoring those processes and
interrupting them when the linked window is closed, or when the app
exits. Furthermore, the application now ensures that child processes,
spawned from the original, are destroyed too.
- To accompany these changes, the Binaries class had some adjustments to
how executions were called and handled.
- Fixed a potential bug where media commands may still execute within a
window that is closing or closed.