Addressed subtle race condition creating invisible and unusable windows.
Builds on prior work and stabilization done in 61d3bb4.

The window was being shown before setting its size and location.
Although both setVisible and setSize operate on the EDT, setVisible
triggers native peer creation, which runs asynchronously. This caused a
race between that native code from setVisible and the setSize call.
Since setVisible does not block until the native window is fully
realized, setSize could execute after the new peer used the old 0x0
dimensions. If the timing was right, and under the right load, the race
resulted in invisible and unusable windows. To prevent issues like this,
the window should be fully configured prior to calling setVisible.