Added a shortcut for sending a window behind other PiP windows.
Also, added a shortcut for bringing a window to the front, but that
should almost never be necessary. Focusing a window alone is enough to
bring it to the front.

For now, the focus order is not preserved. The window that holds focus
afterwards will be the one that was created last (most recently).